move update to baseEnemy class
