+++ Back-To-School/Assets/BackToSchool/Scripts/Enemies/BaseEnemy.cs
@@ -71,6 +71,8 @@ namespace Assets.BackToSchool.Scripts.Enemies
             _agent.SetDestination(newDestination);
         }
 
+        protected virtual void Attack() => _animator.SetTrigger(Constants.AnimationStates.Attack.ToString());
+
         private void Awake()
         {
             _animator = GetComponent<Animator>();
@@ -83,6 +85,62 @@ namespace Assets.BackToSchool.Scripts.Enemies
             _enemyDamage   = EnemyStats.Damage.GetValue();
             _agent.speed   = EnemyStats.MoveSpeed.GetValue();
             _state         = Constants.EnemyStates.Patrolling;
+        }
+
+        private void Update()
+        {
+            if (!_isBusy && !_isDead && _target)
+            {
+                switch (_state)
+                {
+                    case Constants.EnemyStates.Patrolling:
+                        if (!SpaceOperations.CheckIfTwoObjectsClose(transform.position, _target.transform.position, _startChasingDistance))
+                        {
+                            if (SpaceOperations.CheckIfTwoObjectsClose(transform.position, _agent.destination,
+                                _agent.stoppingDistance))
+                                MoveToNextPatrolPoint();
+                        }
+                        else
+                        {
+                            _state = Constants.EnemyStates.Chasing;
+                            _agent.SetDestination(_target.transform.position);
+                        }
+
+                        break;
+                    case Constants.EnemyStates.Chasing:
+                        if (SpaceOperations.CheckIfTwoObjectsClose(transform.position, _agent.destination,
+                            _agent.stoppingDistance))
+                        {
+                            _state = Constants.EnemyStates.Attacking;
+                            _animator.SetBool(Constants.AnimationStates.IsMoving.ToString(), false);
+                        }
+                        else
+                        {
+                            if (!SpaceOperations.CheckIfTwoObjectsClose(transform.position, _target.transform.position,
+                                _startChasingDistance))
+                                _state = Constants.EnemyStates.Patrolling;
+                            else if (_target.transform.position != _agent.destination)
+                                _agent.SetDestination(_target.transform.position);
+                        }
+
+                        break;
+                    case Constants.EnemyStates.Attacking:
+                        if (!SpaceOperations.CheckIfTwoObjectsClose(transform.position, _target.transform.position,
+                            _agent.stoppingDistance))
+                        {
+                            _agent.SetDestination(_target.transform.position);
+                            _state = Constants.EnemyStates.Chasing;
+                            _animator.SetBool(Constants.AnimationStates.IsMoving.ToString(), true);
+                        }
+                        else
+                        {
+                            _isBusy = true;
+                            Attack();
+                        }
+
+                        break;
+                }
+            }
         }
     }
 }
