--- Back-To-School/Assets/BackToSchool/Scripts/Interfaces/IAttackable.cs
@@ -1,7 +1,0 @@
-﻿namespace Assets.BackToSchool.Scripts.Interfaces
-{
-    public interface IAttackable
-    {
-        public void Attack(float damage);
-    }
-}
