+++ Back-To-School/Assets/BackToSchool/Scripts/Interfaces/IMeleeAttackable.cs
@@ -1,0 +1,7 @@
+﻿namespace Assets.BackToSchool.Scripts.Interfaces
+{
+    public interface IMeleeAttackable
+    {
+        public void DoDamage(float damage);
+    }
+}
