+++ Back-To-School/Assets/BackToSchool/Scripts/Interfaces/IMeleeAttackable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 25421a192c26cc543a67cc64a4a1e476
+guid: 66a95faf82fb2fc42aa0b54b245f325e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
